requirements.txt
django
graphene_django
auth0-python
django_filter
django-cors-headers
pyjwt
